+++ bitstring/__init__.py
@@ -17,9 +17,8 @@ pack -- Create a Bits object from a format string.
 
 Exceptions:
 
-Error -- Module exception base class.
-ReadError -- Reading or peeking past the end of a bitstring. Subclasses Error and IndexError.
-ByteAlignError -- Whole byte position or length needed. Subclasses Error.
+ReadError -- Reading or peeking past the end of a bitstring. Subclasses IndexError.
+ByteAlignError -- Whole byte position or length needed. Subclasses ValueError.
 CreationError -- Error during creation. An alias for ValueError.
 InterpretError -- Inappropriate interpretation of binary data. An alias for ValueError.
 
@@ -61,7 +60,7 @@ from .bitarray_ import BitArray
 from .reader import Reader
 from .methods import pack
 from .array_ import Array
-from .exceptions import Error, ReadError, InterpretError, ByteAlignError, CreationError
+from .exceptions import ReadError, InterpretError, ByteAlignError, CreationError
 from .dtypes import DtypeDefinition as _DtypeDefinition, dtype_register as _dtype_register, Dtype
 from typing import Literal as _Literal
 
@@ -299,5 +298,5 @@ if Bits.__doc__ is not None:
 if BitArray.__doc__ is not None:
     BitArray.__doc__ = BitArray.__doc__.replace('[GENERATED_PROPERTY_DESCRIPTIONS]', _property_docstring)
 __all__ = ['Reader', 'BitArray', 'Array',
-           'Bits', 'pack', 'Error', 'ReadError', 'InterpretError',
+           'Bits', 'pack', 'ReadError', 'InterpretError',
            'ByteAlignError', 'CreationError', 'Dtype']